mysite/requirements.txt
asgiref==3.2.9
Django==3.0.7
django-appconf==1.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-nested-admin==3.3.2
django-redis==4.12.1
django-select2==7.4.2
django-tinymce==3.0.2
python-monkey-business==1.0.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
